third_party/WebKit/LayoutTests/inspector/file-system-mapping-overrides-expected.txt

Tests FileSystemMapping overrides


Running: testFileSystemClashDirectOrder
http://localhost:1234/right_url/file.txt

Running: testFileSystemClashReversedOrder
http://localhost:1234/right_url/file.txt

Running: testNetworkClashDirectOrder
file:///Source/devtools/right/file.txt

Running: testNetworkClashReversedOrder
file:///Source/devtools/right/file.txt
